- KtActualWithTODOsMakerVisitorVoid: see property's delegate] - KtActualWithTODOsMakerVisitorVoid: #getFileWithActualsWithTODOs - if declaration is resolvable and can be expect -> delete it, else - leave - KtActualMakerVisitorVoid: #getFileWithActualsWithTODOs - if declaration is resolvable and can be expect -> delete it, else - leave - KtExplicitTypeSpecifierVisitor: remove #getTypeBySuper and replaced with returnType in descriptor of property/fun

+++ idea/idea-gradle/tests/org/jetbrains/kotlin/mppconverter/visitor/KtActualMakerVisitorVoid.kt
@@ -63,7 +63,8 @@ private fun KtDeclaration.makeActual() {
 fun KtFile.getFileWithActuals(): KtFile = (this.copy() as KtFile).apply {
     for (declaration in declarations) {
         if (declaration.isResolvable()) {
-            declaration.delete()
+            if (declaration.canConvertToCommon()) // remove duplicates with common. This way let leave private declarations
+                declaration.delete()
         } else {
             if (declaration.canConvertToCommon())
                 declaration.makeActual()
